Rename files to bothan
Because that’s what it’s called now
