+++ config.ru
@@ -4,7 +4,7 @@ Bundler.setup
 
 ENV['RACK_ENV'] ||= 'development'
 
-require File.join(File.dirname(__FILE__), 'lib/metrics-api.rb')
+require File.join(File.dirname(__FILE__), 'lib/bothan.rb')
 
 use(Rack::Cors) do
   allow do
