+++ features/support/env.rb
@@ -10,7 +10,7 @@ ENV['RACK_ENV'] = 'test'
 ENV['METRICS_API_USERNAME'] = 'foo'
 ENV['METRICS_API_PASSWORD'] = 'bar'
 
-require File.join(File.dirname(__FILE__), '..', '..', 'lib/metrics-api.rb')
+require File.join(File.dirname(__FILE__), '..', '..', 'lib/bothan.rb')
 
 require 'capybara'
 require 'capybara/cucumber'
