+++ lib/bothan/helpers.rb
